Add files via upload
That's a .py file. It's code checks symbols on picture, that you give to it and shows  the largest of all. Now it will show some mistakes, connected with fillPoly. Be sure, i will fix it. Or I will give you another file, that works correct but without one small thing.