Create SolidColor.py
This script can find the largest colorful symbol in the image, that you give him. 
What's an idea:
1. It will find all contours, that bounds your picture.
2. It will find max width*height of all bounding rectangles in your picture.
3. It will create a mask with white background and filled contour of the largest symbol.
4. Script will add two images(mask and source image) by using cv2.add() function.
5. It will give you the image-result.
Please, pay an attention to one thing - now it can't exactly detect color of symbols, so in line 39, please, in a tuple give exact color of symbols. Also don't forget, that it will show it in BGR-format.
Thanks for attention.
Good luck!